Find a way to run shellcheck on all scripts in the project
This took a bit of digging but found a way to use find to return all shell scripts, checking for a shebang with sh, bash or ksh.
I then had great difficulty getting the results of this in a way that I could call the function I had already created to check an individual script.